Refactor email signup endpoint into router
